backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Mail
Flask-CORS
Flask-Migrate
PyMySQL
Flask-Login
Flask-Jwt-Extended
python-dotenv
Flask-session
cryptography
torch
transformers
unsloth
huggingface_hub
accelerate
scikit-learn
lightgbm
pandas
joblib
numpy